version2 adaptations & reconfigs

+++ src/main/java/dat/backend/model/entities/Bottom.java
@@ -3,18 +3,17 @@ package dat.backend.model.entities;
 public class Bottom {
     private int bottomID;
     private String bottoms;
-    private double bottomPrice;
+    private int bottomPrice;
 
-    public Bottom(int bottomID, String bottoms, double bottomPrice) {
+    public Bottom(int bottomID, String bottoms, int bottomPrice) {
         this.bottomID = bottomID;
         this.bottoms = bottoms;
         this.bottomPrice = bottomPrice;
     }
 
-    public Bottom(String bottoms, double bottomPrice) {
+    public Bottom(String bottoms, int bottomPrice) {
         this.bottoms = bottoms;
         this.bottomPrice = bottomPrice;
-
     }
 
     public int getBottomID() {
@@ -29,11 +28,11 @@ public class Bottom {
         this.bottoms = bottoms;
     }
 
-    public double getBottomPrice() {
+    public int getBottomPrice() {
         return bottomPrice;
     }
 
-    public void setBottomPrice(double bottomPrice) {
+    public void setBottomPrice(int bottomPrice) {
         this.bottomPrice = bottomPrice;
     }
 
@@ -45,5 +44,3 @@ public class Bottom {
                 '}';
     }
 }
-
-
